Revert "publish-5 메인페이지"
